fixing according to feedback
